Move era presets into their own component
